+++ src/test/java/com/sparta/swaglabstesting/webdrivers/FirefoxDriverManager.java
@@ -14,7 +14,7 @@ public class FirefoxDriverManager extends WebDriverManager{
             Properties prop = new Properties();
             FileInputStream input = new FileInputStream("src/test/resources/WebDriver.properties");
             prop.load(input);
-            System.setProperty("webdriver.gecko.driver", prop.getProperty("webDriverPath"));
+            System.setProperty("webdriver.gecko.driver", prop.getProperty("FirefoxDriverPath"));
         } catch (IOException e) {
             e.printStackTrace();
         }
